requirements.txt
python-dotenv==0.19.2
Flask==2.0.2
PyMySQL==1.0.2
SQLAlchemy==1.4.28
Flask-Login==0.6.2
Bcrypt-Flask==1.0.2
Flask-WTF==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Faker==10.0.0
Flask-DebugToolbar==0.11.0
google-cloud-storage==2.1.0
Pillow==8.4.0
Flask-Caching==1.10.1
redis==4.1.4
email-validator==1.1.3
flask-paginate==2021.10.29
pandas==1.5.0
openpyxl==3.0.10
xlrd2==1.3.4
pytz==2022.2.1
Flask-Admin==1.6.0